functions/function_app.py
import azure.functions as func
import logging
import json
import os
from typing import Dict, Any

# Initialize the Function App
app = func.FunctionApp(http_auth_level=func.AuthLevel.FUNCTION)

@app.route(route="health", methods=["GET"])
def health_check(req: func.HttpRequest) -> func.HttpResponse:
    """Health check endpoint for Azure Functions"""
    logging.info('Health check requested.')
    
    return func.HttpResponse(
        json.dumps({
            "status": "healthy",
            "service": "ai-doctor-functions",
            "version": "1.0.0"
        }),
        status_code=200,
        mimetype="application/json"
    )

@app.route(route="process-lesion", methods=["POST"])
def process_lesion_analysis(req: func.HttpRequest) -> func.HttpResponse:
    """Process lesion image for AI analysis"""
    logging.info('Processing lesion analysis request.')
    
    try:
        # Get request data
        req_body = req.get_json()
        
        if not req_body:
            return func.HttpResponse(
                json.dumps({"error": "Request body is required"}),
                status_code=400,
                mimetype="application/json"
            )
        
        image_url = req_body.get('image_url')
        patient_id = req_body.get('patient_id')
        
        if not image_url or not patient_id:
            return func.HttpResponse(
                json.dumps({"error": "image_url and patient_id are required"}),
                status_code=400,
                mimetype="application/json"
            )
        
        # TODO: Implement AI model inference
        # This would typically involve:
        # 1. Download image from blob storage
        # 2. Run AI model inference
        # 3. Store results in database
        # 4. Send notification to doctor
        
        analysis_result = {
            "patient_id": patient_id,
            "image_url": image_url,
            "analysis": {
                "confidence": 0.85,
                "classification": "benign",
                "risk_level": "low",
                "recommendations": [
                    "Monitor for changes",
                    "Follow up in 6 months"
                ]
            },
            "processed_at": "2025-09-15T12:00:00Z"
        }
        
        return func.HttpResponse(
            json.dumps(analysis_result),
            status_code=200,
            mimetype="application/json"
        )
        
    except Exception as e:
        logging.error(f"Error processing lesion analysis: {str(e)}")
        return func.HttpResponse(
            json.dumps({"error": "Internal server error"}),
            status_code=500,
            mimetype="application/json"
        )

@app.route(route="send-notification", methods=["POST"])
def send_notification(req: func.HttpRequest) -> func.HttpResponse:
    """Send notification to patients/doctors"""
    logging.info('Processing notification request.')
    
    try:
        req_body = req.get_json()
        
        if not req_body:
            return func.HttpResponse(
                json.dumps({"error": "Request body is required"}),
                status_code=400,
                mimetype="application/json"
            )
        
        notification_type = req_body.get('type')
        recipient_id = req_body.get('recipient_id')
        message = req_body.get('message')
        
        if not all([notification_type, recipient_id, message]):
            return func.HttpResponse(
                json.dumps({"error": "type, recipient_id, and message are required"}),
                status_code=400,
                mimetype="application/json"
            )
        
        # TODO: Implement notification logic
        # This would typically involve:
        # 1. Determine notification method (email, SMS, push)
        # 2. Send notification via appropriate service
        # 3. Log notification in database
        
        result = {
            "notification_id": f"notif_{recipient_id}_{notification_type}",
            "status": "sent",
            "type": notification_type,
            "recipient_id": recipient_id,
            "sent_at": "2025-09-15T12:00:00Z"
        }
        
        return func.HttpResponse(
            json.dumps(result),
            status_code=200,
            mimetype="application/json"
        )
        
    except Exception as e:
        logging.error(f"Error sending notification: {str(e)}")
        return func.HttpResponse(
            json.dumps({"error": "Internal server error"}),
            status_code=500,
            mimetype="application/json"
        )

@app.blob_trigger(arg_name="myblob", path="lesion-images/{name}",
                  connection="AzureWebJobsStorage")
def blob_trigger_lesion_upload(myblob: func.InputStream):
    """Triggered when a new lesion image is uploaded to blob storage"""
    logging.info(f'Blob trigger processed blob: {myblob.name}')
    
    try:
        # TODO: Implement blob processing logic
        # This would typically involve:
        # 1. Extract metadata from blob
        # 2. Queue for AI analysis
        # 3. Update database with upload status
        # 4. Notify relevant parties
        
        logging.info(f'Successfully processed blob: {myblob.name}')
        
    except Exception as e:
        logging.error(f"Error processing blob {myblob.name}: {str(e)}")

@app.timer_trigger(schedule="0 0 2 * * *", arg_name="mytimer", run_on_startup=False)
def daily_cleanup_timer(mytimer: func.TimerRequest) -> None:
    """Daily cleanup task - runs at 2 AM every day"""
    logging.info('Daily cleanup task started.')
    
    try:
        # TODO: Implement cleanup logic
        # This would typically involve:
        # 1. Clean up temporary files
        # 2. Archive old data
        # 3. Update statistics
        # 4. Generate daily reports
        
        logging.info('Daily cleanup task completed successfully.')
        
    except Exception as e:
        logging.error(f"Error in daily cleanup task: {str(e)}")

if __name__ == "__main__":
    # For local development
    logging.basicConfig(level=logging.INFO)
    logging.info("AI Doctor Functions app started locally")