functions/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
requests>=2.31.0
python-multipart>=0.0.6